@@ -2,6 +2,7 @@
   "name": "boxcars",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -10,19 +11,29 @@
   },
   "dependencies": {
     "lucide-react": "^0.544.0",
-    "next": "14.0.4",
+    "next": "15.5.4",
     "next-pwa": "^5.6.0",
-    "react": "^18",
-    "react-dom": "^18"
+    "react": "19.1.1",
+    "react-dom": "19.1.1"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.4.1",
+    "@biomejs/biome": "2.2.4",
+    "@tailwindcss/postcss": "^4.1.13",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "19.1.15",
+    "@types/react-dom": "19.1.9",
     "autoprefixer": "^10.0.1",
-    "postcss": "^8",
-    "tailwindcss": "^3.3.0",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.13",
     "typescript": "^5"
+  },
+  "pnpm": {
+    "overrides": {
+      "@types/react": "19.1.15",
+      "@types/react-dom": "19.1.9"
+    },
+    "onlyBuiltDependencies": [
+      "@tailwindcss/oxide"
+    ]
   }
 }
