@@ -1,4 +1,6 @@
-const withPWA = require("next-pwa")({
+import withPWA from "next-pwa";
+
+const withPWAConfig = withPWA({
   dest: "public",
   disable: process.env.NODE_ENV === "development",
   register: true,
@@ -8,4 +10,4 @@ const withPWA = require("next-pwa")({
 /** @type {import('next').NextConfig} */
 const nextConfig = {};
 
-module.exports = withPWA(nextConfig);
+export default withPWAConfig(nextConfig);
